+++ Website/flaskr/gestione_utente/GestioneUtenteService.py
@@ -1,5 +1,6 @@
 from Website.flaskr.model.Apicoltore import Apicoltore
 from Website.flaskr.model.Cliente import Cliente
+from .. import db
 
 
 def getApicoltoreByEmail(email):
@@ -8,3 +9,7 @@ def getApicoltoreByEmail(email):
 
 def getClienteByEmail(email):
     return Cliente.query.filter_by(email=email).first()
+
+def registraApicoltore(utente):
+    db.session.add(utente)
+    db.session.commit()
