--- Website/flaskr/gestione_utente/RegistrazioneApicoltoreService.py
@@ -1,4 +1,0 @@
-from Website.flaskr.model.Apicoltore import Apicoltore
-
-def getApicoltoreByEmail(email):
-    return Apicoltore.query.filter_by(email=email).first()
