+++ Website/flaskr/templates/base.html
@@ -27,7 +27,7 @@
       <li class="dropdown"><a href="#" class="dropdown-toggle" data-toggle="dropdown"><span class="glyphicon glyphicon-user"></span>Area Utente</a>
        <ul class="dropdown-menu">
           <li><a href="#">Login</a></li>
-          <li><a href="#">Registrati</a></li>
+          <li><a href="{{url_for('views.sigup_ap')}}">Registra apicoltore</a></li>
         </ul></li>
     </ul>
   </div>
