+++ Website/flaskr/templates/registrazione_apicoltore.html
@@ -5,19 +5,19 @@
         document.addEventListener("DOMContentLoaded", function() {
 
 var registrazione = document.createElement('div');
-registrazione.innerHTML = '<form action="action_page.php">' +
+registrazione.innerHTML = '<form action="/registrazione_ap" method="post">' +
   '<div class="container">' +
   '<h1>Registrazione</h1>' +
   '<p>Compila i campi per completare la registrazione.</p>' +
   '<hr>' +
   '<label for="email"><b>Email</b></label>' +
   '<input type="text" placeholder="Email" name="email" required>' +
-  '<label for="psw"><b>Password</b></label>' +
-  '<input type="password" placeholder="Password" name="psw" required>' +
+  '<label for="password"><b>Password</b></label>' +
+  '<input type="password" placeholder="Password" name="password" required>' +
   '<label for="cpwd"><b>Ripeti Password</b></label>' +
   '<input type="password" placeholder="Ripeti Password" name="cpwd" required>' +
   '<label for="nome"><b>Nome</b></label>' +
-  '<input type="text" placeholder="Nome" nome="nome" required>' +
+  '<input type="text" placeholder="Nome" name="nome" required>' +
   '<label for="cognome"><b>Cognome</b></label>' +
   '<input type="text" placeholder="Cognome" name="cognome" required>' +
   '<label for="citta"><b>Citt&#224</b></label>' +
@@ -26,8 +26,8 @@ registrazione.innerHTML = '<form action="action_page.php">' +
   '<input type="text" placeholder="Indirizzo" name="indirizzo" required>' +
   '<label for="cap"><b>Cap</b></label>' +
   '<input type="text" placeholder="Cap" name="cap" required>' +
-  '<label for="numtelefono"><b>Numero di Telefono</b></label>' +
-  '<input type="text" placeholder="Numero di telefono" name="numtelefono" required>' +
+  '<label for="telefono"><b>Numero di Telefono</b></label>' +
+  '<input type="text" placeholder="Numero di telefono" name="telefono" required>' +
   '<label for="assistenza"><b>Assistenza</b></label>' +
   '<input type="text" placeholder="Assistenza" name="assistenza" required>' +
   '<label for="descrizione"><b>Descrizione</b></label>' +
